requirements.txt
fastapi
uvicorn
pydantic
python-multipart
sentence-transformers
faiss-cpu
mlflow
pdfplumber
pytest
groq